Add an Hadoop CDH element
Add an Hadoop CDH element that installs only HDFS from CDH 4.5
Comments in the 40-setup-hadoop script indicate where code should be
added to support other CDH versions and add full support for the
distribution. If this version will be accepted, a bug report will be
created to track the missing features.

This element will be used to create an image with Spark, suitable for
use with the Savanna Spark plugin.

Change-Id: I6202e6f5d88cc697af53ff1fd5d23d35c80b2b6d