Add a Spark element
Add a Spark element that installs Spark in the image. By default the
script tries to find a suitable Spark version based on the Hadoop
version string. For now only the Hadoop CDH installed by the hadoop-cdh
element is supported.
Environment variables can be used to select a different Spark version or
use a custom Spark distribution, generated by hand.

This element is going to be used to build images that will be configured and
run by the Savanna Spark plugin.

Change-Id: Icd4dfef792a7c112bdcccc8951f7e86892abbe86