Update hadoop-cdh element
Update and cleanup the hadoop-cdh element. Depend on a more generic CDH
version, otherwise when the Cloudera repository is updated, packages
installed by the element no longer reflect the version it claims to be
installing.
Use the Oracle Java version provided by the new java element.
Remove the firstboot hack.

Change-Id: If94997d9330d2aee0c23f58051a262b516c2c8e5