Fix a removed variable, properly check the Hadoop version
With the introduction of support for 2.3 and 2.4 the variable
HADOOP_V2_NATIVE_LIBS_DOWNLOAD_URL has been split into
the HADOOP_V2_3_... and HADOOP_V2_4_... variant; this part
of the code needs to be updated as well, otherwise the
installation of native libs fails.

In addition, properly fail if the requested Hadoop version
does not match a know one.

Change-Id: If05afc90a5f081030f7efa449f8bae5b7565a74e