Refactoring IMAGE_TYPE to BASE_IMAGE_OS
Using IMAGE_TYPE to reflect the OS of the base image
conflicts with the way that the diskimage-builder project
uses IMAGE_TYPE (qcow2 vs tar) and was confusing.
We are now using BASE_IMAGE_OS since it properly
reflects what we are using it for and does not conflict
with the diskimage-builder project.

Change-Id: I9e84077c1bc9b8890076838e34a157cb7a3884db
Closes-Bug: #1378364