Adding new update option to diskimage-create.sh
This change removes the global system updates in favor of only updating
the required packages for diskimage creation. Additionally the script
will not automatically install the updates but will warn the user if the
packages are not installed. A switch is provided to allow the script to
install the updates.

Changes
* adding function to detect an installed package
* adding function to check for a list of installed packages
* adding command line switch for updating (-u)
* removing global update commands
* refactoring update logic to fail gracefully or accept switch
* adding -u switch to build-images gate tool
* adding note in documentation about update command line switch

Closes-Bug: 1375843
Change-Id: Iba5e676fa8ef7a852b71acd9e4c19d32f848285a