mapr: use disable-firewall only on centos
The disable-firewall element is specific to centos/fedora, as it refers
to service names available only in centos/fedora.

As mapr can be built on ubuntu as well, remove disable-firewall from
the element-deps (as it makes it always used) and specify it among the
elements used when building a mapr-based image on centos.

Change-Id: Id167b48ad6c97ec141de0f0c17f61d2654ec78ae