Start caching large-sized resources
Use the cache-url script (in the element cache-url) to download and
cache resources which might be expensive (mostly because of their size)
to fetch every time.
As the shared cache ($DIB_IMAGE_CACHE) is available only when running
the root.d elements, move the download phases to root.d scripts.

Change-Id: Iec3e0f92e62c4c9542487a3c228ba8f9e884e5dd