Use the 'base' element implicitly everywhere
Now that the 'centos' element is used for CentOS images, there is no
more need to exclude the 'base' element; thus, the usage of -n for
disk-image-create can be removed.  Also, 'base' makes sure the distro is
up-to-date, so the usage of 'updater' can be removed too.

As side change, since 'base' is implicit unless -n is passed, remove its
usage, also removing the -n usage with it.

Change-Id: I5b829264ac1b3d51a7e3c5fa9b60676c5dcc2e74