fixing minor error in spark post-install script
This change fixes an error with variable substitution in the 20-spark
post install script. The here document firstboot script is using a
substitution for the removal command that was instead substituting the
name of the element script.

This change also adds some minor improvements to reduce repeated code
and make the script more streamlined for future improvement.

Change-Id: I5164cd759fd36c9ad1f910b7801a69cef6873d75
Closes-Bug: 1468131