hadoop: consolidate /run creation
Move the creation of /run (or /var/run if the former does not exist) to
the firstboot script for both hadoop v1 and v2, instead of creating it
with firstboot for v1 and statically in build for v2.

This avoids the same code done in multiple places, and at different
stages.

Change-Id: I2f9461d5435f441814f17c4767958264dc438f6f