hadoop: mark our rc.local replacement as executable
On SysV and upstart rc.local is shipped as executable, but it seems to
be executed regardless, so this just makes our rc.local as the distro
one.

On systemd, though, rc.local is a compatibility script, and to run it
must be executable; hence, make it +x so it can run.

Change-Id: I62bf72f6afb0b58b7552ae94331a3347443ef0af