hadoop: add vanilla/2.6 based on CentOS 7
Add the support for generating vanilla images of Hadoop 2.6 based on
CentOS 7, using the cloud image provided on centos.org (which seems to
work well).

The base element for this is 'centos7', and 'centos7' as $DISTRO_NAME
has been added in the elements used by hadoop; main differences with
CentOS 6 are:
- MariaDB is used instead of MySQL, so install it explicitly (as the
  mysql compatibility names seems not provided); also, adapt
  start-mysql.sh to the new service name
- OpenJDK 8 is the default one, so install that version instead of v7
- iptables-services is needed to have the iptables & ip6tables init
  services

Support for Hadoop v1 is not provided, as the support for Hadoop v1 in
Sahara will be deprecated soon.

Change-Id: I452ae5f935276f58450a4d7c48241fb413271bf1