gate: custom handling of cloudera images
Generating cloudera images requires lots of space, in terms of cached
resources (distribution base images, packages, etc) and generated
images. This causes the generation of all the cloudera images to
(almost?) always fail on the gate slaves, as running out of space.

As a workaround to this, manually invoke the generation of cloudera
images per-distro, cleaning the cache (which contains mostly stuff
specific to the last built distribution) and the generated images after
each run.

Change-Id: I4468f9e2926e101864e73f8aedf7b0612a3f7a96