diskimage-create.sh: add and use image_create
Introduce an helper function to wrap the disk-image-create invocation,
so it is possible to factor out all the common elements, arguments, and
environment variables for all the images.

Current common elements are:
- vm & ntp (already used for all the images)
- xfs-tools (previously not used in ambari images, but should not be an
  huge overhead)
- sahara-version (so all the images have the version information of
  dib-utils, diskimage-builder, and sahara-image-elements)

image_create also does mirror and debug handling for all the images, and
takes care of setting the centos6 cloud image.

Apply image_create instead of invoking disk-image-create, cleaning up
the list of elements, and the environment variables set.

Change-Id: If0e94193383cfa0f4d793353f469ffc4b5d37cab