[CDH] Support CDH 5.7
Added support of CDH 5.7 for ubuntu, centos and centos7.
Also added support of CDH minor versions (5.7.0 and 5.7.1) that we
can set with DIB_CDH_MINOR_VERSION variable
By default it builds 5.7.0 version.

note: for Spark plugin we use HDFS part of CDH so we don't need to
create symbolic links for oozie (so for this we check variable
DIB_CDH_HDFS_ONLY)

depends-on: I1167d0d98ae6fb6fabaf7f1f9a344691d459b50b
Change-Id: I1d71ffbdf78373d27f7e1304164b32da786ac10b
bp: cdh-5-7-support