Fix (again) the detection of scala version (now https)
The www.scala-lang.org switched to https and the curl call did not
account for this scenario, so it failed.
Switch the address to https and as extra measure add also the -L
argument, which forces curl to follow the redirects.

Switch also the download URL to https.

Story: 2001964
Task: 15089
Change-Id: I82d3dd0bd000284fb5404483bf3ee34a6fbb147b